requirements.txt
aiogram==3.1.0
python-dotenv
openai>=1.0.0
google-generativeai>=0.3.2
